requirements.txt
# Core packages
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# PyTorch and related packages
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
torchtext>=0.10.0

# Deep Learning frameworks and utilities
tensorboard>=2.6.0
pytorch-lightning>=1.4.0
fastai>=2.5.0

# Image processing
Pillow>=8.3.0
opencv-python>=4.5.0

# Data manipulation and analysis
scikit-learn>=0.24.0
xgboost>=1.4.0
lightgbm>=3.2.0

# Natural Language Processing
nltk>=3.6.0
spacy>=3.1.0
transformers>=4.9.0

# Data augmentation
albumentations>=1.0.0

# Visualization
plotly>=5.1.0
bokeh>=2.3.0

# Jupyter environment
jupyter>=1.0.0
ipython>=7.26.0

# Model interpretation
shap>=0.39.0
lime>=0.2.0

# Hyperparameter tuning
optuna>=2.9.0
ray[tune]>=1.5.0

# Experiment tracking
mlflow>=1.19.0
wandb>=0.11.0

# Deployment
flask>=2.0.0
fastapi>=0.68.0

# Other useful utilities
tqdm>=4.62.0
pytest>=6.2.0
black>=21.7b0
isort>=5.9.0
mypy>=0.910

# Extended PyTorch ecosystem
torchmetrics>=0.5.0
pytorch3d>=0.6.0
torchdiffeq>=0.2.2
kornia>=0.5.0

# Deep Learning model architectures
timm>=0.4.12

# Audio processing
librosa>=0.8.0

# Video processing
moviepy>=1.0.3

# Reinforcement Learning
gym>=0.19.0
stable-baselines3>=1.2.0

# Genetic Algorithms
deap>=1.3.1

# Graph Neural Networks
torch-geometric>=2.0.0
networkx>=2.6.0

# Probabilistic programming
pyro-ppl>=1.7.0

# Bayesian Deep Learning
pymc3>=3.11.0

# Distributed computing
dask>=2021.7.0